[ft-user-resources] add login authentication
